Version 1.0. Some bugfixes after testing.

+++ debug.h
@@ -1,13 +1,13 @@
 #include <vector>
 
-void out(vector <int> x) {
+void out(vector <double> x) {
 	for (UINT i = 0; i < x.size(); ++i) {
 		cout << ' ' << x[i];
 	}
 	cout << endl;
 }
 
-void output(vector<vector<int> > x) {
+void output(vector<vector<double> > x) {
 	for (UINT i = 0; i < sellers.size(); ++i) {
 		cout << ' ' << sellers[i];
 	}
